change int to uint16 for adc channel

+++ ReadADC/ReadADC.cpp
@@ -4,7 +4,7 @@
 
 
 // constructor
-ADCSensor::ADCSensor(int chip_select, int ADC_channel_number, int zero_mV, int mV_per_sensor_unit){
+ADCSensor::ADCSensor(int chip_select, uint16_t ADC_channel_number, int zero_mV, int mV_per_sensor_unit){
   this->chip_select_ = chip_select;
   this->ADC_channel_number_ = ADC_channel_number;
   this->zero_mV_ = zero_mV;
